+++ app/src/main/java/com/alexvit/android/popularmovies/data/models/api/VideoListResponse.java
@@ -14,7 +14,8 @@ public class VideoListResponse {
 
     @SerializedName("id")
     @Expose
-    public Integer id;
+    public Long id;
+
     @SerializedName("results")
     @Expose
     public List<Video> videos = null;
